irmin: use Odoc @inline annotation to hide aliases into s.ml
Uses the `@inline` annotation to expand references to the signatures
file in the documentation, avoiding unneccesary noise. This can't be
extended to module types that are _not_ aliased in `irmin.ml`, but this
mostly applies to those in `Private` anyway.

Also inlines the inclusion of `Irmin_type` into `Irmin.Type`, hopefully
encouraging users to use the latter.