Extend the implementation of size precomputation for binary codecs
The implementation now makes a distinction between three states of
knowledge about the size of a codec:

- `Static n`: the encoding always occupies exactly `n` bytes;
- `Dynamic f`: the encoding has variable size, pre-computed by `f`;
- `Unknown`: the encoding size cannot be efficiently pre-computed.

This allows some of the work in determining the size of a codec to be
computed at specialisation time, cutting down significantly on the
allocation cost.

Additionally, the implementation is now also able to efficiently
_recover_ the size of an encoding given a pointer to the start of that
encoding (i.e. without actually decoding the value). In this case,
`Dynamic f` is a function that decodes the size of an encoding, and
`Unknown` signifies that this cannot be done (e.g. for unboxed values).

Fix https://github.com/mirage/repr/issues/65.