Revert "Implement Boxed annotation"
This reverts commit b09dddf23f02d52b7d6d4168bc23ef26ef87882e.

Implementing boxing in terms of annotations is conceptually correct, but
interacts with an existing bug in the definition of `pre_hash` (see
https://github.com/mirage/repr/issues/39) that makes this change
breaking inside Irmin.

Pending a solution to https://github.com/mirage/repr/issues/39, we
should avoid changing the behaviour of `pre_hash` in this way. This
restores compatibility with Irmin `master`.